+++ controllers/diseasesCategory.controller.js
@@ -1,6 +1,7 @@
 import DiseasesCategory from "../models/diseasesCategory.model.js";
 import { errorHandler } from "../helpers/errorHandler.js";
 import mongoose from "mongoose";
+import { invalidateCache } from "../helpers/invalidateCache.js";
 
 export const index = async (req, res, next) => {
 	try {
@@ -16,14 +17,15 @@ export const index = async (req, res, next) => {
 			.lean(); // Convert to plain JavaScript objects
 
 		if (diseasesCategories.length === 0) {
-			return next(errorHandler(204, "There aren't any disease categories")); // 204 No Content for empty page
+			return next(
+				errorHandler(204, "There aren't any disease categories")
+			); // 204 No Content for empty page
 		}
 
-		
 		const totalDiseaseCategories = await DiseasesCategory.countDocuments();
 		const totalPages = Math.ceil(totalDiseaseCategories / limit);
 
-		// Return the response 
+		// Return the response
 		return res.status(200).json({
 			data: diseasesCategories,
 			msg: "All disease categories retrieved successfully",
@@ -42,7 +44,6 @@ export const index = async (req, res, next) => {
 		);
 	}
 };
-
 
 export const show = async (req, res, next) => {
 	try {
@@ -71,25 +72,38 @@ export const show = async (req, res, next) => {
 };
 
 export const store = async (req, res, next) => {
-	const { name, description, rank } = req.body;
-	if (name == null || description == null || rank == null) {
-		return next(errorHandler(400, "All required fields must be provided."));
-	}
-	if (description.length > 400) {
-		return next(
-			errorHandler(
-				422,
-				'The field "description" exceeds the maximum length of 400 characters.'
-			)
-		);
+	const user = req.user;
+
+	if (user.role !== "Admin" && user.role !== "Doctor") {
+		return res.status(403).json({
+			success: false,
+			message: "Un-authorized operation",
+		});
 	}
 	try {
+		const { name, description, rank } = req.body;
+		if (name == null || description == null || rank == null) {
+			return next(
+				errorHandler(400, "All required fields must be provided.")
+			);
+		}
+		if (description.length > 400) {
+			return next(
+				errorHandler(
+					422,
+					'The field "description" exceeds the maximum length of 400 characters.'
+				)
+			);
+		}
 		const newDiesasesCategory = await new DiseasesCategory({
 			name,
 			description,
 			rank,
 		});
 		const result = await newDiesasesCategory.save();
+
+		await invalidateCache([`/api/diseasescategories/`]);
+
 		return res.status(201).json({
 			data: result,
 			msg: "New Diesasescategory has been created successfully",
@@ -113,11 +127,19 @@ export const store = async (req, res, next) => {
 };
 
 export const update = async (req, res, next) => {
-	const { id } = req.params;
-	if (id == null) {
-		return next(errorHandler(400, "The 'id' parameter is required."));
+	const user = req.user;
+
+	if (user.role !== "Admin" && user.role !== "Doctor") {
+		return res.status(403).json({
+			success: false,
+			message: "Un-authorized operation",
+		});
 	}
 	try {
+		const { id } = req.params;
+		if (id == null) {
+			return next(errorHandler(400, "The 'id' parameter is required."));
+		}
 		const result = await DiseasesCategory.findOneAndUpdate(
 			{
 				_id: id,
@@ -125,6 +147,10 @@ export const update = async (req, res, next) => {
 			req.body,
 			{ new: true }
 		);
+		await invalidateCache([
+			`/api/diseasescategories/`,
+			`/api/diseasescategories/${id}`,
+		]);
 		return res.status(200).json({
 			data: result,
 			msg: "The DiseasesCategories has successfully updated",
@@ -142,18 +168,34 @@ export const update = async (req, res, next) => {
 };
 
 export const destroy = async (req, res, next) => {
-	const { id } = req.params;
-	if (id == null) {
-		return next(errorHandler(400, "All required fields must be provided."));
+	const user = req.user;
+
+	if (user.role !== "Admin" && user.role !== "Doctor") {
+		return res.status(403).json({
+			success: false,
+			message: "Un-authorized operation",
+		});
 	}
 	try {
+		const { id } = req.params;
+		if (id == null) {
+			return next(
+				errorHandler(400, "All required fields must be provided.")
+			);
+		}
 		const diseasesCategories = await DiseasesCategory.findById(id);
 		if (!diseasesCategories) {
-			return res.status(404).json({ message: "DiseasesCategories not found" });
+			return res
+				.status(404)
+				.json({ message: "DiseasesCategories not found" });
 		}
-		diesasescategory = await DiseasesCategory.findOneAndDelete({
+		await DiseasesCategory.findOneAndDelete({
 			_id: id,
 		});
+		await invalidateCache([
+			`/api/diseasescategories/`,
+			`/api/diseasescategories/${id}`,
+		]);
 		return res.status(204).json({
 			msg: "The DiseasesCategory has been successfully deleted",
 			success: true,
